refactor to merge host and remote clients
