+++ Dockerfile
@@ -11,6 +11,8 @@ RUN npm install --production
 RUN npm install -g typescript
 ARG REACT_APP_URL=https://vidqr.fly.dev
 ARG REACT_APP_API_URL=https://vidqr.fly.dev
+# ARG REACT_APP_URL=http://44.203.74.143
+# ARG REACT_APP_API_URL=http://44.203.74.143
 
 RUN npm run build
