+++ client/src/App.tsx
@@ -4,16 +4,15 @@ import { BrowserRouter as Router, Routes, Route} from 'react-router-dom';
 
 import Home from './pages/Home';
 import Host from './pages/Host';
-import Remote from './pages/Remote';
+import Room from './pages/Room';
 
 const App = () => {
   return (
       <Router>
         <Routes>
           <Route path='/' element={<Home />} />
-          <Route path='/host/:roomId' element={<Host />} />
-          <Route path='/host/' element={<Host />} />
-          <Route path='/:roomId' element={<Remote />}/>
+            <Route path='/host/:roomId' element={<Host />} />
+          <Route path='/:roomId' element={<Room />}/>
         </Routes>
       </Router>
   );
