+++ client/src/components/VideoSearch.tsx
@@ -1,13 +1,8 @@
-import React, { useState, useEffect } from "react";
-import {useNavigate, useParams} from 'react-router-dom';
-import {VideoQueue} from "../components/video_queue";
-import {Button, SimpleGrid, Loader, Stack, TextInput, Pagination} from '@mantine/core';
-import {socket} from "../config/socket";
-import aFetch from "../config/axios";
-import "./remote.css";
-import "./search.css";
-import {ShareLink} from "./Host";
 import {useWindowDimensions} from "../hooks/hooks";
+import React, {useState} from "react";
+import aFetch from "../config/axios";
+import {Button, Loader, Pagination, SimpleGrid, TextInput} from "@mantine/core";
+import {ShareLink} from "../pages/Host";
 
 function VideoIcon(params: {thumbnailLink: string, title: string, channelName: string, videoLink: string, queueVideo: any}) {
 
@@ -45,7 +40,7 @@ function RemoteVideoSearcher(params: {queueVideo: any}) {
         setSubmitted(true);
         setLastQuery(searchQuery);
 
-        aFetch.post('/api/remote/search/', {videoPlatform: 'youtube', query: searchQuery}).then(response => {
+        aFetch.post('/api/room/search/', {videoPlatform: 'youtube', query: searchQuery}).then(response => {
             setVideoResults(response.data);
             setPage(1);
             setSubmitted(false);
@@ -117,10 +112,9 @@ function RemoteVideoSearcher(params: {queueVideo: any}) {
             </div>
         </div>
     )
-
 }
 
-function RemoteWrapper(params: any) {
+function VideoSearcher(params: {roomId: string}) {
     const [videoLink, setVideoLink] = useState('');
     const [submitted, setSubmitted] = useState(false);
     const [invalid, setInvalid] = useState(false);
@@ -128,7 +122,7 @@ function RemoteWrapper(params: any) {
     const textSubmit = () => {
         if (videoLink !== "") {
             setSubmitted(true);
-            aFetch.post(`/api/remote/add_video`, {'roomId': params.roomId, 'videoLink': videoLink}).then(
+            aFetch.post(`/api/room/add_video`, {'roomId': params.roomId, 'videoLink': videoLink}).then(
                 response => {
                     setSubmitted(false);
                     if (response.data.validVideo) {
@@ -142,6 +136,7 @@ function RemoteWrapper(params: any) {
             setInvalid(false);
         }
     }
+
 
     const setVideoLinkAndFocus = (newVideoLink: string) => {
         setInvalid(false);
@@ -161,120 +156,26 @@ function RemoteWrapper(params: any) {
             }
         }
     }
-
     return (
-        <div className="hViewer">
-          <div className="primary">
-              <RemoteVideoSearcher queueVideo={setVideoLinkAndFocus}/>
-              <TextInput
-                  id="videoLinkInput"
-                  placeholder="https://www.youtube.com/watch?v=dQw4w9WgXcQ"
-                  value={videoLink}
-                  onChange={e => {setInvalid(false);setVideoLink(e.target.value)}}
-                  onKeyDown={(e) => e.key === "Enter" ? textSubmit(): null}
-                  rightSection={submitted ? <Loader size={"xs"}/>: null}
-                  error={invalid ? "Invalid Video Link": null}
-              />
-              <Button variant="gradient" gradient={{ from: 'indigo', to: 'cyan' }} onClick={() => {textSubmit()}}>Queue Video</Button>
-              <ShareLink link={params.roomId}/>
-          </div>
-          <div className="secondary">
-            <VideoQueue roomId={params.roomId} username={params.username}/>
-          </div>
-        </div>
-    );
+        <>
+            <RemoteVideoSearcher queueVideo={setVideoLinkAndFocus}/><TextInput
+            id="videoLinkInput"
+            placeholder="https://www.youtube.com/watch?v=dQw4w9WgXcQ"
+            value={videoLink}
+            onChange={e => {
+                setInvalid(false);
+                setVideoLink(e.target.value);
+            }}
+            onKeyDown={(e) => e.key === "Enter" ? textSubmit() : null}
+            rightSection={submitted ? <Loader size={"xs"}/> : null}
+            error={invalid ? "Invalid Video Link" : null}/><Button variant="gradient"
+                                                                   gradient={{from: 'indigo', to: 'cyan'}}
+                                                                   onClick={() => {
+                                                                       textSubmit();
+                                                                   }}>Queue Video</Button><ShareLink
+            link={params.roomId}/>
+        </>
+    )
 }
-const Remote = () => {
-    let params = useParams();
-    let navigate = useNavigate();
 
-    const [username, setUsername] = useState(undefined)
-    const [usernameBox, setUsernameBox] = useState('')
-    const [roomInfo, setRoomInfo] = useState({'roomName': undefined});
-
-    const [error, setError] = useState('');
-
-    useEffect(() => {
-        aFetch.post(`/api/host/get_room_info/`, {'roomId': params.roomId}).then(response => {
-            if (response.data.roomName){
-                document.title = response.data.roomName;
-                setRoomInfo(response.data);
-            }else{
-                navigate('/');
-            }
-        });
-
-        aFetch.post(`/api/remote/get_username/${params.roomId}`).then(response => {
-            if (response.data.username){
-                setUsername(response.data.username);
-            }
-
-        }
-    )}, [navigate, params.roomId]);
-
-    const promptSubmit = () => {
-        if (usernameBox.length > 16){
-            return;
-        }
-
-
-        aFetch.post(`/api/remote/join_room/${params.roomId}`, {'redirect': false, 'username': usernameBox}).then(response => {
-            if (response.data.username){
-                socket.emit("video:subscribe", {'roomId': response.data.roomId});
-                setUsername(response.data.username);
-            }else if(response.data.validRoom){
-                setError('That username already exists.');
-            }
-        });
-
-    }
-
-    const usernameEdit = (username: string) => {
-        setUsernameBox(username);
-        if (username.length > 16){
-            setError('Usernames must be at most 16 characters');
-        }else{
-            setError('');
-        }
-    }
-
-    if (username === '' || username === undefined){
-        return (
-            <div className="mainViewer"><div className="promptForm">
-                <div className="promptBox">
-                    <Stack >
-                        <b>You are trying to join room: <br/><mark>{roomInfo.roomName}</mark></b>
-
-                        <b>Pick a username</b>
-                        <TextInput
-                            placeholder="Username"
-                            onChange={(e) => {usernameEdit(e.target.value)}}
-                            onKeyDown={(e) => e.key === "Enter" ? promptSubmit(): null}
-                            error={error==='' ? null: error}
-                        />
-                        {params.isLocked ? (
-                            <><b>This room is password-locked</b><TextInput
-                                placeholder="Room Secret"/></>): null}
-                        <Button variant="gradient" gradient={{ from: 'teal', to: 'cyan' }} onClick={promptSubmit}>Enter</Button>
-                    </Stack>
-                </div>
-            </div>
-            </div>
-        )
-    }else{
-        return (
-            <div className="mainViewer">
-                <div className="remoteHeader">
-                    <b>{username}</b>
-
-                    <b>{params.roomId}</b>
-                </div>
-
-
-                <RemoteWrapper roomId={params.roomId} username={username}/>
-            </div>
-        );
-    }
-};
-
-export default Remote;
+export default VideoSearcher;
