+++ client/src/pages/Host.tsx
@@ -9,68 +9,17 @@ import {
     CopyButton,
     Tooltip,
     ActionIcon,
-    Grid, Stack, TextInput, Button
 } from '@mantine/core';
 import {IconCopy, IconCheck, } from '@tabler/icons-react';
 import {useParams, useNavigate} from "react-router-dom";
 import {socket} from "../config/socket";
 import {CURRENT_URL} from "../config/url";
 import aFetch from "../config/axios";
-import "./host.css";
-
-function HostVideoPlayer(params: any) {
-    const [videoRef, setVideoRef] = useState('');
-    const [videoPlaying, setVideoPlaying] = useState(true);
-
-    const nextVideo = (discard?: boolean) => {
-        socket.emit("video:nextVideo", {"roomId": params.link, "discard": discard});
-    }
+import "./Host.css";
+import {PromptBox, RoomNamePrompt, RoomUserNamePrompt} from "../components/PromptBox";
+import VideoPlayer from "../components/VideoPlayer";
 
 
-    aFetch.post('/api/host/get_current_video/'+params.link).then(response => {
-        if (response.data.video){
-            setVideoRef(response.data.video.videoLink)
-        }else{
-            setVideoRef('')
-        }
-    })
-
-    socket.on("video:nextVideo", (params: any) => {
-        const videoLink = params.videoLink;
-
-        if (videoLink !== videoRef) {
-            setVideoRef(videoLink);
-        }else{
-            setVideoRef(videoLink + '?');
-        }
-        setVideoPlaying(true);
-    });
-
-
-
-    socket.on("video:toggleVideo", () => {
-        console.log("toggle video");
-        setVideoPlaying(!videoPlaying);
-    });
-
-
-
-    return (
-        <AspectRatio ratio={16 / 9}>
-            <ReactPlayer url={videoRef}
-                         playing={videoPlaying}
-                         onStart={() => setVideoPlaying(true)}
-                         onPause={() => setVideoPlaying(false)}
-                         onError={() => new Promise(res => setTimeout(res, 2000)).then(() => nextVideo(true))}
-                         controls={true}
-                         embedoptions={{cc_load_policy: 1, cc_lang_pref: "en"}}
-                         width="100%"
-                         height="100%"
-                         onEnded={() => nextVideo()}
-            />
-        </AspectRatio>
-    );
-}
 class UserList extends React.Component<{}, {data: TransferListData}>  {
   constructor(props: any) {
     super(props);
@@ -160,7 +109,7 @@ function HostMenu(props: { roomId: string; }) {
             <div className="hViewer">
                 <div className="vViewer">
                     <div className="primary">
-                        <HostVideoPlayer link={props.roomId}/>
+                        <VideoPlayer link={props.roomId}/>
                     </div>
                     <div className="hostSettings">
                         <ShareLink link={props.roomId}/>
@@ -181,10 +130,11 @@ const Host = () => {
 
     const [roomName, setRoomName] = useState('');
     const [roomNameBox, setRoomNameBox] = useState('');
+    const [roomNameError, setRoomNameError] = useState('');
 
     useEffect(() => {
         if (params.roomId === "create_room"){
-            aFetch.post('/api/host/rejoin_room').then(response => {
+            aFetch.post('/api/room/rejoin_room').then(response => {
                 if (response.data.host){
                     navigate(`/host/${response.data.roomId}`);
                 }
@@ -195,7 +145,7 @@ const Host = () => {
         }else if (invalidRoomId) {
             navigate('/host/create_room', {replace: true});
         }else{
-            aFetch.post(`/api/host/get_room_info`, {'roomId': params.roomId}).then(response => {
+            aFetch.post(`/api/room/get_room_info`, {'roomId': params.roomId}).then(response => {
                 if (response.data.host){
                     setRoomName(response.data.roomName);
                 }else{
@@ -205,43 +155,33 @@ const Host = () => {
         }
     }, [navigate, params.roomId, invalidRoomId]);
 
+
     const promptSubmit = () => {
-        if (roomNameBox.length > 16){
+        if (roomNameBox.length > 16 || roomNameBox.length === 0){
+            if (roomNameBox.length === 0){
+                setRoomNameError('Room name cannot be empty');
+            }
             return;
         }
 
-        aFetch.post(`/api/host/create_room/`, {'roomName': roomNameBox }).then(response => {
+        aFetch.post(`/api/room/create_room/`, {'roomName': roomNameBox }).then(response => {
             if (response.data.roomId){
                 setRoomName(roomNameBox);
-                navigate(`/host/${response.data.roomId}`)
+                navigate(`/${response.data.roomId}`)
             }
         });
-
     }
 
     if (params.roomId === 'create_room'){
         return (
-            <div className="mainViewer">
-                <div className="promptForm">
-                    <div className="promptBox">
-                        <Stack >
-                            <b>Pick a room name</b>
-                            <TextInput
-                                placeholder="Room Name"
-                                onChange={(e) => {setRoomNameBox(e.target.value)}}
-                                onKeyDown={(e) => e.key === "Enter" ? promptSubmit(): null}
-                                error={roomNameBox.length > 16 ? "Room names must be at most 16 characters": null}
-                            />
-                            {/*<b>Optional Password</b>*/}
-                            {/*<TextInput*/}
-                            {/*    placeholder="Password"*/}
-                            {/*    onChange={(e) => {}}*/}
-                            {/*/>*/}
-                            <Button variant="gradient" gradient={{ from: 'teal', to: 'cyan' }} onClick={promptSubmit}>Enter</Button>
-                        </Stack>
-                    </div>
-                </div>
-            </div>
+            <PromptBox promptSubmit={promptSubmit}>
+                <RoomNamePrompt roomNameState={[roomNameBox, setRoomNameBox]} errorState={[roomNameError, setRoomNameError]}/>
+                {/*<b>Optional Password</b>*/}
+                {/*<TextInput*/}
+                {/*    placeholder="Password"*/}
+                {/*    onChange={(e) => {}}*/}
+                {/*/>*/}
+            </PromptBox>
         )
     }else if (invalidRoomId || params.roomId === undefined) {
         return (
@@ -250,18 +190,13 @@ const Host = () => {
     }else{
         document.title = roomName;
         return (
-            <>
-
-                <div className="mainViewer">
-                    <div className="remoteHeader">
-                        <b>{roomName}</b>
-
-                        <b>{params.roomId}</b>
-                    </div>
-                    <HostMenu roomId={params.roomId}/>
+            <div className="mainViewer">
+                <div className="remoteHeader">
+                    <b>{roomName}</b>
+                    <b>{params.roomId}</b>
                 </div>
-            </>
-
+                <HostMenu roomId={params.roomId}/>
+            </div>
         );
     }
 };
